Reorganize, basic service/controller change
Reorganized the app structure to help separate controllers, services,
dtos, models, etc. This should make the app easier to navigate as it
starts to get bigger.
In addition, added some simple service/interface/controller code to see
how the dependency injection built in to Nest.js works. Also, the power
of interfaces is incredible! I love that I don't even have to state the
DTO for CreateCat implements the Cat interface for it to be passed to
the CatService.